requirements.txt
tensorflow
pillow
numpy
fastapi[standard]
requests
